pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "plumtree-agents"
version = "0.1.0"
description = "Internal agents for Plumtree Services"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "python-docx>=1.1.0",
    "slack-sdk>=3.27.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
]

[project.scripts]
plumtree-research = "agents.research.agent:main"
plumtree-proposal = "agents.proposal_sow.agent:main"
plumtree = "main:main"

[tool.setuptools.packages.find]
include = ["agents*", "config*", "integrations*", "shared*"]
